feat: Refactor MediaInner to utilize MediaDetail and RelatedPreview components
